Bi Network Connection Loop
The BiNetworkConnections.cs script is a connection loop between two
devices a mobile and a computer, being that in unity3d without using the
Network Manager System, the client cannot make changes throughout the
server. So this script is like a ping poog game between the two devices
only that the ball is the connection. Note that the script is not
complete yet.
The MobilexPCServer.cs script is a code that would find the two devices
and start the connection loop.